pyproject.toml
[project]
name = "languize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "sqlalchemy-orm>=1.2.10",
    "uvicorn>=0.35.0",
]
